chore(): refactor types

+++ pkg/models/product.go
@@ -4,9 +4,9 @@ import "gorm.io/gorm"
 
 type Product struct {
 	gorm.Model
-	Id                int32            `json:"id" gorm:"primaryKey"`
+	Id                int64            `json:"id" gorm:"primaryKey"`
 	Name              string           `json:"name"`
-	Stock             int32            `json:"stock"`
-	Price             int32            `json:"price"`
+	Stock             int64            `json:"stock"`
+	Price             int64            `json:"price"`
 	StockDecreaseLogs StockDecreaseLog `gorm:"foreignKey:ProductRefer"`
 }
